requirements.txt
# Core Dependencies
fastapi==0.103.0
uvicorn==0.23.2
streamlit==1.25.0
python-dotenv==1.0.0

# Data & Analysis
pandas
numpy
yfinance==0.2.28

# API and WebSocket
websockets==11.0.3
websocket-client
aiohttp
requests

# Visualization
plotly==5.15.0
matplotlib
seaborn